fix: Prevent integer overflow in MaxSegmentBytes calculation
When MaxSegmentMs=300000 (5 minutes), the calculation
16000 * 300000 = 4.8 billion overflows int32 (max ~2.1 billion),
resulting in ~1MB limit instead of the expected 9.6MB.

Fixed by casting to long before multiplication.

+++ src/VirtualAssistant.Core/Configuration/ContinuousListenerOptions.cs
@@ -88,7 +88,7 @@ public class ContinuousListenerOptions
     // Computed properties
     public int ChunkSizeBytes => SampleRate * VadChunkMs / 1000 * 2; // 16-bit = 2 bytes per sample
     public int PreBufferMaxBytes => SampleRate * PreBufferMs / 1000 * 2;
-    public int MaxSegmentBytes => SampleRate * MaxSegmentMs / 1000 * 2; // Max audio size for transcription
+    public int MaxSegmentBytes => (int)((long)SampleRate * MaxSegmentMs / 1000 * 2); // Max audio size for transcription (use long to prevent overflow)
 
     /// <summary>
     /// Gets the full path for WhisperModelPath, resolving relative paths.
